Deleted Test.vue, migrated content to Portfolio.vue. Changed images & fixed tile layout
